+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # unreleased
-* Forward `forwardedComments` to `processorComments`
+* Add `processorComments` to Dispute
 
 # 4.3.0
 * Add `PayPalHereDetails` to transactions